requirements.txt
discord.py==2.0.0
yt-dlp==2023.3.4
spotipy==2.19.0
python-dotenv==1.0.0
